allow specifying a pre-computed X-Amz-Content-Sha256 value
We're looking to use this library in a distributed environment where our server generates a list of generic http requests (e.g. method, url and headers) which get passed to a remote machine which does the actual uploading.  Our remote machine shouldn't have any aws secrets, so we need to generate the list of all these http request's metadata on the server.  Because the server will never have access to the actual content, we need to compute the value of the sha256 for the request body, send that to the server and sign based on that value.

This quick patch allows us to do that by skipping the hashing of the body.  It would also make it easier to use this library to do streaming uploads of things to s3 since it would allow you to use the streams api to compute the sha256 hash, do the signing, then use streams to write to the aws request.